+++ Sources/Flare/Classes/Models/StoreProduct.swift
@@ -61,7 +61,7 @@ extension StoreProduct: ISKProduct {
     public var currencyCode: String? {
         product.currencyCode
     }
-    
+
     public var currencySymbol: String? {
         product.currencySymbol
     }